+++ CppProjectFiles/Cell.h
@@ -7,15 +7,15 @@ enum CellType {REGULAR, FORREST, SEA, FLAG_A, FLAG_B};
 
 class Cell {
 	static int aliveIns;
-    size_t   row ;
-    size_t   column;
+    uint   row ;
+    uint   column;
     CellType type;
     Ship*    standingShip = nullptr;
 
     void setCellType(CellType aType) { type = aType; }
     
 public:
-    Cell(size_t row, size_t column, CellType type = REGULAR)
+    Cell(uint row, uint column, CellType type = REGULAR)
         : row(row), column(column), type(type)  {
 		aliveIns++;
 	}
